Create CSV class
This class contains two methods, initializing CSV objects and importing CSV files. Importing a CSV file requires a clear path and CSV file name. The CSV data is imported row by row and stored in an array. Each row of the imported data can be retrieved by reading the elements in the array.