Add Data File Control Scripts
Initialize the "FileController.cs". "FileController.cs" is used to control the actual data file in the game by associating it with the Game Object in the game. FileController.cs calls the methods of the CSV class, creates a CSV instance, imports the datafile, and reads the datafile line by line.